warnings fixed

+++ dune-stokes/src/dune_stokes.cc
@@ -302,7 +302,7 @@ void AccuracyRun( CollectiveCommunication& mpicomm )
     // setting this to true will give each run a unique logilfe name
     bool per_run_log_target = Parameters().getParam( "per-run-log-target", true );
 
-    int numruns = std::pow( accurracy_steps, 2.0 );
+	int numruns = int( std::pow( accurracy_steps, 2.0 ) );
     profiler().Reset( numruns );
     for ( int i = 0; i < accurracy_steps; ++i ) {
         for ( int j = 0; j < accurracy_steps; ++j ) {
@@ -392,7 +392,7 @@ void StabRun( CollectiveCommunication& mpicomm )
 
     int ref = Parameters().getParam( "minref", 0 );
 
-    // setting this to true will give each run a unique logilfe name
+	// setting this to true will give each run a unique logfile name
     bool per_run_log_target = Parameters().getParam( "per-run-log-target", true );
 
     CoeffVector coeff_vector = getC_power_Permutations();
@@ -446,7 +446,7 @@ void BfgRun( CollectiveCommunication& mpicomm )
     const double stop_tau = Parameters().getParam( "bfg-tau-stop", 0.4 ) ;
     const double tau_inc = Parameters().getParam( "bfg-tau-inc", 0.025 ) ;
 
-    profiler().Reset( ( stop_tau - start_tau ) / tau_inc + 1 );
+	profiler().Reset( int( ( stop_tau - start_tau ) / tau_inc + 1 ) );
 
     for ( double tau = start_tau; tau < stop_tau; tau += tau_inc ) {
         Parameters().setParam( "bfg-tau", tau );
